Take of async and  await to simplify nested promises calls

+++ server/server.js
@@ -55,19 +55,18 @@ app.get('/todos/:id', authenticate, (req, res) => {
   }).catch(e => res.status(400).send());
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-  let id = req.params.id;
-  if (!ObjectID.isValid(id)) return res.status(404).send();
-
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
+  try {
+    let id = req.params.id;
+    if (!ObjectID.isValid(id)) throw new Error(404);
+    let todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id });
     if (!todo) {
-      return res.status(404).send();
+      throw new Error(404);
     }
     res.send({ todo });
-  }).catch(e => res.status(400).send());
+  } catch (e) {
+    res.status(e.message).send();
+  }
 });
 
 app.patch('/todos/:id', authenticate, (req, res) => {
@@ -90,45 +89,50 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 
 });
 
-app.post('/users', (req, res) => {
-  let body = _.pick(req.body, ['email', 'password']);
-  let user = new User(body);
-
-  user.save()
-    .then(() => {
-      return user.generateAuthToken();
-    })
-    .then((token) => {
-      res.header('x-auth', token).send(user);
-    })
-    .catch((e) => {
-      res.status(400).send(e);
-    });
+app.post('/users', async (req, res) => {
+  try {
+    let body = _.pick(req.body, ['email', 'password']);
+    let user = new User(body);
+    await user.save();
+    let token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
+    res.status(400).send(e);
+  }
+  // user.save()
+  //   .then(() => {
+  //     return user.generateAuthToken();
+  //   })
+  //   .then((token) => {
+  //     res.header('x-auth', token).send(user);
+  //   })
+  //   .catch((e) => {
+  //     res.status(400).send(e);
+  //   });
 });
 
 app.get('/users/me', authenticate, (req, res) => {
   res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-  let { email, password } = _.pick(req.body, ['email', 'password']);
-  User.findByCredentials(email, password)
-    .then((doc) => {
-      return doc.generateAuthToken().then((token) => {
-        res.header('x-auth', token).send(doc);
-      });
-    })
-    .catch((e) => {
-      res.status(404).send(e);
-    });
+app.post('/users/login', async (req, res) => {
+  try {
+    let { email, password } = _.pick(req.body, ['email', 'password']);
+    let user = await User.findByCredentials(email, password);
+    let token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
+    res.status(404).send(e);
+  }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-  req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+  try {
+    await req.user.removeToken(req.token);
     res.status(200).send();
-  }, (e) => {
+  } catch (e) {
     res.status(400).send();
-  })
+  }
 });
 
 app.listen(port, () => {
